+++ pints/__init__.py
@@ -237,18 +237,18 @@ from ._nested._ellipsoid import NestedEllipsoidSampler
 
 
 #
-# Sampling initialising
-#
-from ._sample_initial_points import sample_initial_points
-
-#
 # ABC
 #
-
 from ._abc import ABCSampler
 from ._abc import ABCController
 from ._abc._abc_rejection import RejectionABC
 from ._abc._abc_smc import ABCSMC
+
+
+#
+# Sampling initialising
+#
+from ._sample_initial_points import sample_initial_points
 
 
 #