+++ pints/toy/__init__.py
@@ -32,4 +32,3 @@ from ._sho_model import SimpleHarmonicOscillatorModel
 from ._simple_egg_box import SimpleEggBoxLogPDF
 from ._sir_model import SIRModel
 from ._twisted_gaussian_banana import TwistedGaussianLogPDF
-from ._stochastic_logistic_model import StochasticLogisticModel